+++ CHANGELOG.md
@@ -18,6 +18,10 @@ For a full diff see [`0.7.0...main`][0.7.0...main].
 - Change visual representation of tags in table view to match Obsidian's style
 - Automatically focus search field in library view
 
+### Fixed
+
+- Trim settings to prevent unintended spaces from slipping through
+
 ## [`0.7.0`][0.7.0] - 2025-04-02
 
 For a full diff see [`0.6.1...0.7.0`][0.6.1...0.7.0].