+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-For a full diff see [`0.7.0...main`][0.7.0...main].
+For a full diff see [`0.8.0...main`][0.8.0...main].
+
+## [`0.8.0`][0.8.0] - 2025-04-03
+
+For a full diff see [`0.7.0...0.8.0`][0.7.0...0.8.0].
 
 ### Added
 
@@ -114,6 +118,7 @@ For a full diff see [`0.2.2...0.3.0`][0.2.2...0.3.0].
 [0.6.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.6.0
 [0.6.1]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.6.1
 [0.7.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.7.0
+[0.8.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.8.0
 [0.2.2...0.3.0]: https://github.com/weph/obsidian-bookshelf/compare/0.2.2...0.3.0
 [0.3.0...0.3.1]: https://github.com/weph/obsidian-bookshelf/compare/0.3.0...0.3.1
 [0.3.1...0.4.0]: https://github.com/weph/obsidian-bookshelf/compare/0.3.1...0.4.0
@@ -121,4 +126,5 @@ For a full diff see [`0.2.2...0.3.0`][0.2.2...0.3.0].
 [0.5.0...0.6.0]: https://github.com/weph/obsidian-bookshelf/compare/0.5.0...0.6.0
 [0.6.0...0.6.1]: https://github.com/weph/obsidian-bookshelf/compare/0.6.0...0.6.1
 [0.6.1...0.7.0]: https://github.com/weph/obsidian-bookshelf/compare/0.6.1...0.7.0
-[0.7.0...main]: https://github.com/weph/obsidian-bookshelf/compare/0.7.0...main
+[0.7.0...0.8.0]: https://github.com/weph/obsidian-bookshelf/compare/0.7.0...0.8.0
+[0.8.0...main]: https://github.com/weph/obsidian-bookshelf/compare/0.8.0...main